The worker directory has been restructured to include multiple directories.
