+++ worker/task/load/model_loader.py
@@ -5,11 +5,12 @@ import os
 import time
 import json
 
-from .task_response import TaskResponse
-from .llm_model import LLMModel
+
+from ...task_response import TaskResponse
+from ...llm_model import LLMModel
 from schemas import ModelLoadParams
 
-from .logger_config import setup_logger
+from ...logger_config import setup_logger
 log_level = os.environ.get("LOG_LEVEL", "INFO")
 logger = setup_logger(__name__, level=log_level)
