+++ worker/task/token-count/tokenizer_service.py
@@ -3,10 +3,10 @@ from transformers import PreTrainedTokenizer
 import os
 import json
 
-from .task_response import TaskResponse
-from .llm_model import LLMModel
+from ...task_response import TaskResponse
+from ...llm_model import LLMModel
 from schemas import TokenCountParams
-from .logger_config import setup_logger
+from ...logger_config import setup_logger
 
 log_level = os.environ.get("LOG_LEVEL", "INFO")
 logger = setup_logger(__name__, level=log_level)
